move 508 support functions into cypress support configuration
